Switches from overide to new injection class
