+++ lib/hyku_addons/engine.rb
@@ -468,7 +468,6 @@ module HykuAddons
       ::ApplicationController.include HykuAddons::MultitenantLocaleControllerBehavior
       ::Hyku::API::V1::FilesController.include HykuAddons::FilesControllerBehavior
       ::ActiveJob::Base.include HykuAddons::ImportMode
-      Hyrax::Hirmeos::WorkFactory.prepend HykuAddons::WorkFactoryBehavior
     end
 
     # Use #to_prepare because it reloads where after_initialize only runs once
